set up main tabs navigator
