+++ src/app/(tabs)/debug/index.tsx
@@ -1,0 +1,8 @@
+import { View, Text, StyleSheet } from 'react-native';
+import DebugScreen from 'screens/DebugScreen';
+
+// export default function Index() {
+//     return (
+//         <DebugScreen/>
+//     );
+// }
