+++ src/app/(tabs)/game.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import GameScreen from 'screens/GameScreen';
+
+const GamePage: React.FC = () => {
+    return (
+        <GameScreen/>
+    );
+};
+
+export default GamePage;
