+++ src/app/(tabs)/index.tsx
@@ -1,0 +1,8 @@
+import { View, Text, StyleSheet } from 'react-native';
+import HomeScreen from 'screens/HomeScreen';
+
+export default function Index() {
+    return (
+        <HomeScreen/>
+    );
+}
