+++ src/app/(tabs)/insights.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import PlaceholderScreen from 'screens/PlaceholderScreen';
+
+const Insights: React.FC = () => {
+    return <PlaceholderScreen />;
+};
+
+export default Insights;
