+++ src/screens/GameScreen.tsx
@@ -7,7 +7,7 @@ import { View, StyleSheet } from 'react-native';
 import Game from 'components/game_components/Game'; // Adjust the import path if needed
 import { GameState, NewGameState } from 'components/game_components/Game';
 
-const GameScreen = ({ navigation}) => {
+const GameScreen = () => {
     // This is a set of puzzles for testing purposes. 
     const gameStates: GameState[] = [
         NewGameState("#1: Pastry puzzle: Part 1", "I **** to *** **n**k**", "I love to eat pancakes", 4, 1),
