+++ src/utils/CapitalizeString.ts
@@ -1,0 +1,3 @@
+export function capitalize(s: string) {
+    return [s.toLocaleUpperCase()[0], s.slice(1)].join('');
+}
